snapshot-20001120

+++ postfix/src/global/mail_print.c
@@ -150,7 +150,7 @@ int     mail_vprint(VSTREAM *stream, const char *fmt, va_list ap)
 	    continue;
 	if (*cp != '%')
 	    msg_fatal("mail_vprint: bad format: %.*s>%c<%s",
-		      cp - fmt, fmt, *cp, cp + 1);
+		      (int) (cp - fmt), fmt, *cp, cp + 1);
 	if ((lflag = (*++cp == 'l')) != 0)
 	    cp++;
 
@@ -186,7 +186,7 @@ int     mail_vprint(VSTREAM *stream, const char *fmt, va_list ap)
 		}
 	    if (tp >= mail_print_tab + mail_print_tablen)
 		msg_fatal("mail_vprint: bad format: %.*s>%c<%s",
-			  cp - fmt, fmt, *cp, cp + 1);
+			  (int) (cp - fmt), fmt, *cp, cp + 1);
 	}
     }
     return (error);
